山东码上后台管理系统(后端部分):表名修改、定时任务方法逻辑调整

+++ ruoyi-management/src/main/java/com/ruoyi/query/pangolinDataReport/DataReportSecondQuery.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class DataReportSecondQuery {
-    private Integer user_id;
-    private Integer role_id;
+    private Long user_id;
+    private Long role_id;
     private Integer timestamp;
     private String version;
     private String date;
